+++ invenio_userprofiles/__init__.py
@@ -22,7 +22,7 @@
 # waive the privileges and immunities granted to it by virtue of its status
 # as an Intergovernmental Organization or submit itself to any jurisdiction.
 
-"""Invenio module that adds userprofiles to the platform."""
+"""User profiles module for Invenio."""
 
 from __future__ import absolute_import, print_function