+++ tests/helpers.py
@@ -22,7 +22,6 @@
 # waive the privileges and immunities granted to it by virtue of its status
 # as an Intergovernmental Organization or submit itself to any jurisdiction.
 
-
 """Helper functions for tests."""
 
 from __future__ import absolute_import, print_function
@@ -30,17 +29,23 @@ from __future__ import absolute_import, print_function
 from flask import url_for
 
 
-def sign_up(app, client):
+def sign_up(app, client, email=None, password=None):
     """Register a user."""
-    client.post(url_for('security.register'), data=dict(
-        email=app.config['TEST_USER_EMAIL'],
-        password=app.config['TEST_USER_PASSWORD'],
+    with app.test_request_context():
+        register_url = url_for('security.register')
+
+    client.post(register_url, data=dict(
+        email=email or app.config['TEST_USER_EMAIL'],
+        password=password or app.config['TEST_USER_PASSWORD'],
     ), environ_base={'REMOTE_ADDR': '127.0.0.1'})
 
 
-def login(app, client):
+def login(app, client, email=None, password=None):
     """Log the user in with the test client."""
-    client.post(url_for('security.login'), data=dict(
-        email=app.config['TEST_USER_EMAIL'],
-        password=app.config['TEST_USER_PASSWORD'],
+    with app.test_request_context():
+        login_url = url_for('security.login')
+
+    client.post(login_url, data=dict(
+        email=email or app.config['TEST_USER_EMAIL'],
+        password=password or app.config['TEST_USER_PASSWORD'],
     ))