+++ tests/test_models.py
@@ -27,7 +27,9 @@
 from __future__ import absolute_import, print_function
 
 import pytest
-
+from invenio_accounts.models import User
+from invenio_db import db
+from sqlalchemy.exc import IntegrityError
 from test_validators import test_usernames
 
 from invenio_userprofiles import UserProfile
@@ -49,3 +51,46 @@ def test_userprofiles(app):
     profile.last_name = 'User'
     assert profile.first_name == 'Test'
     assert profile.last_name == 'User'
+
+
+def test_case_insensitive_username(app):
+    """Test case-insensitive uniqueness."""
+    with app.app_context():
+        with db.session.begin_nested():
+            u1 = User(email='test@example.org')
+            u2 = User(email='test2@example.org')
+            db.session.add(u1, u2)
+        profile1 = UserProfile(user=u1, username="INFO")
+        profile2 = UserProfile(user=u2, username="info")
+        db.session.add(profile1)
+        db.session.add(profile2)
+        pytest.raises(IntegrityError, db.session.commit)
+
+
+def test_case_preserving_username(app):
+    """Test that username preserves the case."""
+    with app.app_context():
+        with db.session.begin_nested():
+            u1 = User(email='test@example.org')
+            db.session.add(u1)
+        db.session.add(UserProfile(user=u1, username="InFo"))
+        db.session.commit()
+        profile = UserProfile.get_by_username('info')
+        assert profile.username == 'InFo'
+
+
+def test_delete_cascade(app):
+    """Test that deletion of user, also removes profile."""
+    with app.app_context():
+        with db.session.begin_nested():
+            u = User(email='test@example.org')
+            db.session.add(u)
+        p = UserProfile(user=u, username="InFo")
+        db.session.add(p)
+        db.session.commit()
+
+        assert UserProfile.get_by_userid(u.id) is not None
+        db.session.delete(u)
+        db.session.commit()
+
+        assert UserProfile.get_by_userid(u.id) is None